Stage-1 pass
