+++ Body.py
@@ -1,0 +1,13 @@
+from Const import (WIDTH, HEIGHT)
+
+
+
+class Body:
+    def __init__(self):
+        self.size = 64
+        self.x = WIDTH / 2 - self.size / 2
+        self.y = HEIGHT - 25 - self.size
+        self.vel_y = 0
+        self.acceleration = 0
+        self.angle = 0
+        self.jumpable = self.vel_y <= 0  # If body is hitting a level, then it can jump only if the body is going down.
