+++ Const.py
@@ -1,0 +1,13 @@
+GAME_FPS = 150
+WIDTH, HEIGHT = 800, 600
+JUMPING_HEIGHT = 20
+MAX_ACCELERATION = 10
+VEL_X = 3  # Setting the moving speed.
+VEL_Y = JUMPING_HEIGHT  # Setting the jumping height.
+
+# Walls settings:
+WALLS_Y = -128
+WALL_WIDTH = 128
+WALLS_ROLLING_SPEED = 2
+RIGHT_WALL_BOUND = WIDTH - WALL_WIDTH
+LEFT_WALL_BOUND = WALL_WIDTH
